+++ src/plugins/lokijs/lokijs-helper.ts
@@ -1,9 +1,14 @@
-import type { RxStorageInstanceLoki } from './rx-storage-instance-loki';
-import type { RxStorageKeyObjectInstanceLoki } from './rx-storage-key-object-instance-loki';
+import { createLokiLocalState, RxStorageInstanceLoki } from './rx-storage-instance-loki';
+import {
+    createLokiKeyValueLocalState,
+    RxStorageKeyObjectInstanceLoki
+} from './rx-storage-key-object-instance-loki';
 import lokijs, { Collection } from 'lokijs';
 import type {
     LokiDatabaseSettings,
     LokiDatabaseState,
+    LokiLocalDatabaseState,
+    LokiRemoteResponseBroadcastMessage,
     MangoQuery,
     MangoQuerySortDirection,
     MangoQuerySortPart,
@@ -12,12 +17,17 @@ import type {
 import {
     add as unloadAdd, AddReturn
 } from 'unload';
-import { flatClone } from '../../util';
+import { ensureNotFalsy, flatClone, promiseWait, randomCouchString } from '../../util';
 import { LokiSaveQueue } from './loki-save-queue';
 import type { DeterministicSortComparator } from 'event-reduce-js';
 import { getPrimaryFieldOfPrimaryKey } from '../../rx-schema';
 import { newRxError } from '../../rx-error';
-import { BroadcastChannel, createLeaderElection, LeaderElector } from 'broadcast-channel';
+import {
+    BroadcastChannel,
+    createLeaderElection,
+    LeaderElector,
+    OnMessageHandler
+} from 'broadcast-channel';
 import type { RxStorageLoki } from './rx-storage-lokijs';
 
 export const CHANGES_COLLECTION_SUFFIX = '-rxdb-changes';
@@ -264,5 +274,208 @@ export function removeLokiLeaderElectorReference(
             electorState.leaderElector.broadcastChannel.close();
             storage.leaderElectorByLokiDbName.delete(databaseName);
         }
+    }
+}
+
+/**
+ * For multi-instance usage, we send requests to the RxStorage
+ * to the current leading instance over the BroadcastChannel.
+ */
+export async function requestRemoteInstance(
+    instance: RxStorageInstanceLoki<any> | RxStorageKeyObjectInstanceLoki,
+    operation: string,
+    params: any[]
+): Promise<any | any[]> {
+    const isRxStorageInstanceLoki = typeof (instance as any).query === 'function';
+    const messageType = isRxStorageInstanceLoki ? LOKI_BROADCAST_CHANNEL_MESSAGE_TYPE : LOKI_KEY_OBJECT_BROADCAST_CHANNEL_MESSAGE_TYPE;
+
+    const leaderElector = ensureNotFalsy(instance.internals.leaderElector);
+    await waitUntilHasLeader(leaderElector);
+    const broadcastChannel = leaderElector.broadcastChannel;
+
+    type WinningPromise = {
+        retry: boolean,
+        result?: any;
+        error?: any;
+    }
+
+    let whenDeathListener: OnMessageHandler<any>;
+    const leaderDeadPromise = new Promise<WinningPromise>(res => {
+        whenDeathListener = (msg: any) => {
+            if (msg.context === 'leader' && msg.action === 'death') {
+                res({
+                    retry: true
+                });
+            }
+        };
+        broadcastChannel.addEventListener('internal', whenDeathListener);
+    });
+    const requestId = randomCouchString(12);
+    let responseListener: OnMessageHandler<any>;
+    const responsePromise = new Promise<WinningPromise>((res, rej) => {
+        responseListener = (msg: any) => {
+            if (
+                msg.type === messageType &&
+                msg.response === true &&
+                msg.requestId === requestId
+            ) {
+                if (msg.isError) {
+                    res({
+                        retry: false,
+                        error: msg.result
+                    });
+                } else {
+                    res({
+                        retry: false,
+                        result: msg.result
+                    });
+                }
+            }
+        };
+        broadcastChannel.addEventListener('message', responseListener);
+    });
+
+    // send out the request to the other instance
+    broadcastChannel.postMessage({
+        response: false,
+        type: messageType,
+        operation,
+        params,
+        requestId,
+        databaseName: instance.databaseName,
+        collectionName: instance.collectionName
+    });
+
+
+    return Promise.race([
+        leaderDeadPromise,
+        responsePromise
+    ]).then(firstResolved => {
+        // clean up listeners
+        broadcastChannel.removeEventListener('message', responseListener);
+        broadcastChannel.removeEventListener('internal', whenDeathListener);
+
+        if (firstResolved.retry) {
+            /**
+             * The leader died while a remote request was running
+             * we re-run the whole operation.
+             * We cannot just re-run requestRemoteInstance()
+             * because the current instance might be the new leader now
+             * and then we have to use the local state instead of requesting the remote.
+             */
+            return (instance as any)[operation](...params);
+        } else {
+            if (firstResolved.error) {
+                throw firstResolved.error;
+            } else {
+                return firstResolved.result;
+            }
+        }
+    });
+}
+
+/**
+ * Handles a request that came from a remote instance via requestRemoteInstance()
+ * Runs the requested operation over the local db instance and sends back the result.
+ */
+export async function handleRemoteRequest(
+    instance: RxStorageInstanceLoki<any> | RxStorageKeyObjectInstanceLoki,
+    msg: any
+) {
+    const isRxStorageInstanceLoki = typeof (instance as any).query === 'function';
+    const messageType = isRxStorageInstanceLoki ? LOKI_BROADCAST_CHANNEL_MESSAGE_TYPE : LOKI_KEY_OBJECT_BROADCAST_CHANNEL_MESSAGE_TYPE;
+
+    if (
+        msg.type === messageType &&
+        msg.requestId &&
+        msg.databaseName === instance.databaseName &&
+        msg.collectionName === instance.collectionName &&
+        !msg.response
+    ) {
+        const operation = (msg as any).operation;
+        const params = (msg as any).params;
+        let result: any;
+        let isError = false;
+        try {
+            result = await (instance as any)[operation](...params);
+        } catch (err) {
+            isError = true;
+            result = err;
+        }
+        const response: LokiRemoteResponseBroadcastMessage = {
+            response: true,
+            requestId: msg.requestId,
+            databaseName: instance.databaseName,
+            collectionName: instance.collectionName,
+            result,
+            isError,
+            type: msg.type
+        };
+        ensureNotFalsy(instance.internals.leaderElector).broadcastChannel.postMessage(response);
+    }
+}
+
+
+export async function waitUntilHasLeader(leaderElector: LeaderElector) {
+    while (
+        !leaderElector.hasLeader
+    ) {
+        await leaderElector.applyOnce();
+        await promiseWait(0);
+    }
+}
+
+/**
+ * If the local state must be used, that one is returned.
+ * Returns false if a remote instance must be used.
+ */
+export async function mustUseLocalState(
+    instance: RxStorageInstanceLoki<any> | RxStorageKeyObjectInstanceLoki
+): Promise<LokiLocalDatabaseState | false> {
+    if (instance.closed) {
+        return false;
+    }
+
+    const isRxStorageInstanceLoki = typeof (instance as any).query === 'function';
+
+    if (instance.internals.localState) {
+        return instance.internals.localState;
+    }
+    const leaderElector = ensureNotFalsy(instance.internals.leaderElector);
+    await waitUntilHasLeader(leaderElector);
+
+    /**
+     * It might already have a localState after the applying
+     * because another subtask also called mustUSeLocalState()
+     */
+    if (instance.internals.localState) {
+        return instance.internals.localState;
+    }
+
+    if (
+        leaderElector.isLeader &&
+        !instance.internals.localState
+    ) {
+        // own is leader, use local instance
+        if (isRxStorageInstanceLoki) {
+            instance.internals.localState = createLokiLocalState<any>({
+                databaseName: instance.databaseName,
+                collectionName: instance.collectionName,
+                options: instance.options,
+                schema: (instance as RxStorageInstanceLoki<any>).schema,
+                multiInstance: instance.internals.leaderElector ? true : false
+            }, instance.databaseSettings);
+        } else {
+            instance.internals.localState = createLokiKeyValueLocalState({
+                databaseName: instance.databaseName,
+                collectionName: instance.collectionName,
+                options: instance.options,
+                multiInstance: instance.internals.leaderElector ? true : false
+            }, instance.databaseSettings);
+        }
+        return ensureNotFalsy(instance.internals.localState);
+    } else {
+        // other is leader, send message to remote leading instance
+        return false;
     }
 }