+++ src/plugins/worker/workers/lokijs-fs.worker.ts
@@ -1,5 +1,5 @@
-import { getRxStorageLoki } from '../../plugins/lokijs';
-import { wrappedRxStorage } from '../../plugins/worker';
+import { getRxStorageLoki } from '../../lokijs';
+import { wrappedRxStorage } from '../../worker';
 
 const lfsa = require('lokijs/src/loki-fs-structured-adapter.js');
 const adapter = new lfsa();