+++ src/plugins/worker/workers/lokijs-incremental-indexeddb.worker.ts
@@ -1,0 +1,10 @@
+import { getRxStorageLoki } from '../../lokijs';
+import { wrappedRxStorage } from '../../worker';
+const LokiIncrementalIndexedDBAdapter = require('lokijs/src/incremental-indexeddb-adapter');
+
+const storage = getRxStorageLoki({
+    adapter: new LokiIncrementalIndexedDBAdapter()
+});
+wrappedRxStorage({
+    storage
+});