+++ src/plugins/worker/workers/lokijs-memory.worker.ts
@@ -1,0 +1,7 @@
+import { getRxStorageLoki } from '../../lokijs';
+import { wrappedRxStorage } from '../../worker';
+
+const storage = getRxStorageLoki();
+wrappedRxStorage({
+    storage
+});