+++ src/plugins/worker/workers/pouchdb-idb.worker.ts
@@ -1,0 +1,10 @@
+import { getRxStoragePouch, addPouchPlugin } from '../../pouchdb';
+import { wrappedRxStorage } from '../../worker';
+
+addPouchPlugin(require('pouchdb-adapter-idb'));
+const storage = getRxStoragePouch({
+    adapter: 'idb'
+});
+wrappedRxStorage({
+    storage
+});